requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
pillow==10.0.1
pydirectinput==1.0.4
numpy==1.24.3
keyboard==0.13.5
mouse==0.7.1
pywin32==306
scipy==1.11.4
scikit-learn==1.3.2 
matplotlib==3.7.2